Re-work error handling in encode and encode, plus perform image-type conversion
Re-work the error handling in Steganography::encode and Steganography::decode into a more readable form, hopefully.

Perform an image-type conversion on the input images to ensure that they are RGBA16. This will avoid potentially creating errors later.

Add some comments to indicate that the test data and key are just for testing purposes. This should be obvious, but just in case!