Refactor encoding method
Refactor the encoding method to make things cleaner and easier to understand.

Instead of selecting a random cell each time data is encoded, the cell map (data index -> cell id) is automatically compiled.

When writing (or reading) a byte of data, the index of the byte is located in the cell map. The value at that index will give the cell ID to be used for the specified byte.