Make DataWrapperV1 hold a ChaCha20Rng of its own
DataWrapperV1 will now hold a ChaCha20Rng instance of its own, to avoid complicating things with lifetimes and sharing it with StegaV1.

The selection system for deciding if a value should be adjusted by ±1 will be done based on the original value instead. The value will be nudged towards 127, which is the median byte value.