Rework the concept of the image wrapper, using bytes instead of pixels
Rework the concept of the image wrapper, using bytes instead of pixels: the wrapper now works with with the raw bytes representing the channels of the pixel, rather than the pixels themselves.

This allows the code to be faster as we can more easily create mutable references to the data, and it also permits the code to be significantly cleaned up.