Faster random byte implementation
Implement a faster method for filling a vector with bytes.

This system works by generating 64-byte blocks of u8 values and then appending those onto the vector. In the even that the total number of entries modulo 64 is not equal to 0, then the remaining entries will be filled via a loop.