Refactor the data decoder
Refactor the data decoder. The decoder now handles the XOR decoding internally, which cleans up the methods and makes things simpler when interacting with the decoder. Once loading and decoding is done then popping values from the internal vector is really simple.

I might make similar changes to the encoder...