Remove version
The byte representing the version number will no longer be encoded into the image file.

Upon reflection it isn't needed. As the algorithm to determine where the cells are located can be changed version-to-version it's not really possible to determine the version of the file by looking at a specific place in the file.

In addition to this, adding more information to help validate if a file contains data completely undermines the point.

Due to these factors the version will no longer be encoded.