Code clean-up and simplification
Some minor code clean-up and simplification.

We do not need to calculate the difference between the reference and encoded images as the difference will always be 0, 1 or -1. This means that we can just check that they are different in order to determine if a bit should be set or not. Naturally this would not be the case if a codec were used that permitted higher degrees of variance.