Hook up the locker to the v1 decoder
The v1 decoder will now correctly check for the zTXt section of a PNG file. If present it will attempt to extract the feature flags from it. The only flag currently supported is the use_locker flag.

While the use_locker functionality is functional, the CLI application does not currently make use of it. There are also currently no tests for the file locker, these will be addressed in another commit soon.