More test refactoring, and a new test
Another refactor of the test system, the file cleaner will be directly build into TestUtils. This leads to simpler code, and I should have thought of it sooner.

Added a new test to StegaV1 that will test if the generation of the composite key remains fixed. While refactoring I accidentally changed the way the key was generated, which rendered any files created prior to that change undecodable. This test should prevent that problem cropping up again.