Refactor the way feature flags are handled, and add some tests
Feature flags are now stored as flags, since it's going to get messy as more are added for various features.

Added a test to ensure that files are encoded with a PNG zTXt, and a test to check whether the file locker feature flag is correctly enabled and disabled based on the data from the zTXt chunk.