Update dependencies
Due to a change with the rand crate, the order of shuffle is no longer consister with the old version. As such this would break prior tests and files encoded with the old system would no longer correctly decode.

I will be adding a v2 codec to address this in the near future, and allow the more modern version of shuffle to be used.