Add the ability for the CLI to "guess" the version of the codec used to encode the file.
Note that, when it reads a value larger than the current maximum version, it will assume it is a v1 file for legacy reasons.

Edited the codecs to no longer validate the version when decoding. This will allow people to tamper with the header (for whateve reason) and still successfully decode the file.